:> working on installation and README
