+++ crystal_building/builder.py
@@ -1,10 +1,10 @@
 import torch
 from crystal_building.utils import \
-    (update_supercell_data, unit_cell_to_convolution_cluster, clean_cell_output,
+    (update_supercell_data, unit_cell_to_convolution_cluster,
      align_crystaldata_to_principal_axes,
      batch_asymmetric_unit_pose_analysis_torch, set_sym_ops,
-     rotvec2rotmat, build_unit_cell, scale_asymmetric_unit)
-from common.geometry_calculations import compute_fractional_transform, sph2rotvec, rotvec2sph
+     rotvec2rotmat, build_unit_cell)
+from common.geometry_calculations import compute_fractional_transform_torch, sph2rotvec
 from constants.asymmetric_units import asym_unit_dict
 
 
@@ -37,19 +37,21 @@ class SupercellBuilder:
         # sort fractional vectors from closest to furthest from central unit cell
         self.sorted_fractional_translations = fractional_translations[torch.argsort(fractional_translations.abs().sum(1))].to(device)
 
-    def build_supercells(self, molecule_data, cell_sample: torch.tensor, supercell_size: int = 5,
-                         graph_convolution_cutoff: float = 6, target_handedness=None,
-                         align_molecules=False, pare_to_convolution_cluster=True):
+    def build_supercells(self,
+                         molecule_data,
+                         cell_sample: torch.tensor,
+                         supercell_size: int = 5,
+                         graph_convolution_cutoff: float = 6,
+                         target_handedness=None,
+                         align_to_standardized_orientation=False,
+                         pare_to_convolution_cluster=True):
         """
         convert cell parameters to unit cell in a fast, differentiable, invertible way
         convert reference cell to "supercell" (in fact, it's truncated to an appropriate cluster size)
         """
         supercell_data = molecule_data.clone()
-
         supercell_data, cell_sample, target_handedness = \
             self.move_cell_data_to_device(supercell_data, cell_sample, target_handedness)
-
-        sym_ops_list, supercell_data = set_sym_ops(supercell_data)  # assign correct symmetry options
 
         cell_lengths, cell_angles, mol_position, mol_rotation_i = (
             cell_sample[:, :3], cell_sample[:, 3:6], cell_sample[:, 6:9], cell_sample[:, 9:])
@@ -60,37 +62,37 @@ class SupercellBuilder:
             mol_rotation = mol_rotation_i
 
         # get transformation matrices
-        T_fc_list, T_cf_list, generated_cell_volumes = compute_fractional_transform(cell_lengths, cell_angles)
-
+        T_fc_list, T_cf_list, generated_cell_volumes = compute_fractional_transform_torch(cell_lengths, cell_angles)
         supercell_data.T_fc = T_fc_list
-        supercell_data.cell_params = torch.cat((cell_lengths, cell_angles, mol_position, mol_rotation), dim=1)  # note: destandardized
+        supercell_data.cell_params = torch.cat((cell_lengths, cell_angles, mol_position, mol_rotation), dim=1)
+        sym_ops_list, supercell_data = set_sym_ops(supercell_data)  # assign correct symmetry options
+        rotations_list = rotvec2rotmat(mol_rotation)
 
-        if align_molecules:  # align canonical conformers principal axes to cartesian axes - not usually done here, but allowed
+        if align_to_standardized_orientation:  # align canonical conformers principal axes to cartesian axes - not usually done here, but allowed
             supercell_data = align_crystaldata_to_principal_axes(supercell_data, handedness=target_handedness)
 
-        # get molecule information
-        coords_list = []
-        atomic_number_list = []
-        for i in range(supercell_data.num_graphs):
-            atomic_number_list.append(supercell_data.x[supercell_data.batch == i])
-            coords_list.append(supercell_data.pos[supercell_data.batch == i])
+        # get molecule information  # todo check this against the below
+        atomic_number_list = [supercell_data.x[supercell_data.batch == i] for i in range(supercell_data.num_graphs)]
+        coords_list = [supercell_data.pos[supercell_data.batch == i] for i in range(supercell_data.num_graphs)]
 
-        # convert rotvecs to rotation matrices
-        rotations_list = rotvec2rotmat(mol_rotation)
+        # atomic_number_list = []
+        # coords_list = []
+        # for i in range(supercell_data.num_graphs):
+        #     atomic_number_list.append(supercell_data.x[supercell_data.batch == i])
+        #     coords_list.append(supercell_data.pos[supercell_data.batch == i])
 
         # center, apply rotation, apply translation (to canonical conformer)
         canonical_conformer_coords_list = []
         for i, (rotation, coords, T_fc, new_frac_pos) in enumerate(zip(rotations_list, coords_list, T_fc_list, mol_position)):
             canonical_conformer_coords_list.append(
-                torch.inner(rotation, coords - coords.mean(0)).T
-                + torch.inner(T_fc, new_frac_pos)
+                torch.inner(rotation, coords - coords.mean(0)).T + torch.inner(T_fc, new_frac_pos)
             )
 
         # apply symmetry ops to build unit cell
         unit_cell_coords_list = build_unit_cell(supercell_data.mult, canonical_conformer_coords_list, T_fc_list, T_cf_list, sym_ops_list)
 
         # reanalyze the constructed unit cell to get the canonical orientation & confirm correct construction
-        mol_positions, mol_orientations, mol_handedness, well_defined_asym_unit = \
+        _, mol_orientations, mol_handedness, _ = \
             batch_asymmetric_unit_pose_analysis_torch(
                 unit_cell_coords_list,
                 supercell_data.sg_ind,
@@ -102,19 +104,19 @@ class SupercellBuilder:
         supercell_data.cell_params[:, 9:12] = mol_orientations  # overwrite to canonical parameters
         supercell_data.asym_unit_handedness = mol_handedness
 
-        cell_vector_list = T_fc_list.permute(0, 2, 1)
-
         # get minimal supercell cluster for convolving about a given canonical conformer
+        cell_vector_list = T_fc_list.permute(0, 2, 1)
         supercell_list, supercell_atoms_list, ref_mol_inds_list, n_copies = \
             unit_cell_to_convolution_cluster(
                 unit_cell_coords_list, cell_vector_list, T_fc_list, atomic_number_list, supercell_data.mult,
-                supercell_scale=supercell_size, cutoff=graph_convolution_cutoff,
+                supercell_scale=supercell_size,
+                cutoff=graph_convolution_cutoff,
                 sorted_fractional_translations=self.sorted_fractional_translations,
                 pare_to_convolution_cluster=pare_to_convolution_cluster)
 
         supercell_data = update_supercell_data(supercell_data, supercell_atoms_list, supercell_list, ref_mol_inds_list, unit_cell_coords_list)
 
-        return supercell_data, generated_cell_volumes, None
+        return supercell_data, generated_cell_volumes
 
     def prebuilt_unit_cell_to_supercell(self, supercell_data, supercell_size=5, graph_convolution_cutoff=6, pare_to_convolution_cluster=True):
         """
@@ -124,9 +126,6 @@ class SupercellBuilder:
         automatically pare NxNxN supercell to minimal set of molecules in the convolution radius of the canonical conformer
         """
         supercell_data = supercell_data.clone().to(self.device)
-
-        # T_fc_list, T_cf_list, generated_cell_volumes = compute_fractional_transform(
-        #     cell_lengths=supercell_data.cell_params[:, 0:3], cell_angles=supercell_data.cell_params[:, 3:6])
 
         atoms_list = []
         for i in range(supercell_data.num_graphs):
@@ -144,44 +143,6 @@ class SupercellBuilder:
         supercell_data = update_supercell_data(supercell_data, supercell_atoms_list, supercell_list, ref_mol_inds_list, supercell_data.ref_cell_pos)
 
         return supercell_data.to(self.device)
-
-    def process_cell_params(self, supercell_data, cell_sample, skip_cell_cleaning=False, standardized_sample=True, rescale_asymmetric_unit=True):
-        if skip_cell_cleaning:  # don't clean up
-            if standardized_sample:
-                destandardized_cell_sample = (cell_sample * torch.tensor(self.dataDims['lattice_stds'], device=self.device, dtype=cell_sample.dtype)) + torch.tensor(
-                    self.dataDims['lattice_means'], device=self.device, dtype=cell_sample.dtype)  # destandardize
-                cell_lengths, cell_angles, mol_position, mol_rotation_i = destandardized_cell_sample.split(3, 1)
-            else:
-                cell_lengths, cell_angles, mol_position, mol_rotation_i = cell_sample.split(3, 1)
-
-            if self.rotation_basis == 'spherical':
-                mol_rotation = sph2rotvec(mol_rotation_i)
-            else:
-                mol_rotation = mol_rotation_i
-        else:
-            cell_lengths, cell_angles, mol_position, mol_rotation = cell_sample.split(3, 1)
-            lattices = [self.symmetries_dict['lattice_type'][int(supercell_data.sg_ind[n])] for n in range(supercell_data.num_graphs)]
-            cell_lengths, cell_angles, mol_position, mol_rotation, _, _, _ = \
-                clean_cell_output(
-                    cell_lengths, cell_angles, mol_position, mol_rotation, lattices,
-                    self.dataDims['lattice_means'], self.dataDims['lattice_stds'],
-                    enforce_crystal_system=True, return_transforms=True,
-                    standardized_sample=standardized_sample, rotation_basis=self.rotation_basis)
-
-        if rescale_asymmetric_unit:
-            mol_position = scale_asymmetric_unit(self.asym_unit_dict, mol_position, supercell_data.sg_ind)
-
-        return cell_lengths, cell_angles, mol_position, mol_rotation
-
-    '''
-    K = torch.stack((
-                torch.stack((torch.zeros_like(unit_vector[:,0]),-unit_vector[:,2], unit_vector[:,1]),dim=1),
-                  torch.stack((unit_vector[:,2], torch.zeros_like(unit_vector[:,0]), -unit_vector[:,0]),dim=1),
-                  torch.stack((-unit_vector[:,1],unit_vector[:,0],torch.zeros_like(unit_vector[:,0])),dim=1)
-                  ),dim=1)
-      
-    R = torch.eye(3)[None,:,:].tile(34,1,1) +torch.sin(theta[:,None,None])*K + (1-torch.cos(theta[:,None,None])) * (K@K)
-    '''
 
     def move_cell_data_to_device(self, supercell_data, cell_sample, target_handedness):
         supercell_data = supercell_data.to(self.device)
