+++ models/basis_functions.py
@@ -1,147 +1,8 @@
-from math import sqrt, pi as PI
+from math import pi as PI
 import torch
-import sympy as sym
-from torch_geometric.nn.models.dimenet_utils import (bessel_basis,
-                                                     associated_legendre_polynomials,
-                                                     sph_harm_prefactor)
 
 
-def real_sph_harm(k, zero_m_only=True, spherical_coordinates=True):
-    if not zero_m_only:
-        S_m = [0]
-        C_m = [1]
-        for i in range(1, k):
-            x = sym.symbols('x')
-            y = sym.symbols('y')
-            S_m += [x * S_m[i - 1] + y * C_m[i - 1]]
-            C_m += [x * C_m[i - 1] - y * S_m[i - 1]]
-
-    P_l_m = associated_legendre_polynomials(k, zero_m_only)
-    if spherical_coordinates:
-        theta = sym.symbols('theta')
-        z = sym.symbols('z')
-        for i in range(len(P_l_m)):
-            for j in range(len(P_l_m[i])):
-                if type(P_l_m[i][j]) != int:
-                    P_l_m[i][j] = P_l_m[i][j].subs(z, sym.cos(theta))
-        if not zero_m_only:
-            phi = sym.symbols('phi')
-            for i in range(1, len(S_m)):
-                S_m[i] = S_m[i].subs(x,
-                                     sym.sin(theta) * sym.cos(phi)).subs(
-                    y,
-                    sym.sin(theta) * sym.sin(phi))
-            for i in range(1, len(C_m)):
-                C_m[i] = C_m[i].subs(x,
-                                     sym.sin(theta) * sym.cos(phi)).subs(
-                    y,
-                    sym.sin(theta) * sym.sin(phi))
-
-    Y_func_l_m = [['0'] * (2 * j + 1) for j in range(k)]
-    for i in range(k):
-        Y_func_l_m[i][0] = sym.simplify(sph_harm_prefactor(i, 0) * P_l_m[i][0])
-
-    if not zero_m_only:
-        for i in range(1, k):
-            for j in range(1, i + 1):
-                Y_func_l_m[i][j] = sym.simplify(
-                    2 ** 0.5 * sph_harm_prefactor(i, j) * C_m[j] * P_l_m[i][j])
-        for i in range(1, k):
-            for j in range(1, i + 1):
-                Y_func_l_m[i][-j] = sym.simplify(
-                    2 ** 0.5 * sph_harm_prefactor(i, -j) * S_m[j] * P_l_m[i][j])
-
-    return Y_func_l_m
-
-
-class SphericalBasisLayer(torch.nn.Module):
-    def __init__(self, num_spherical, num_radial, cutoff=5.0,
-                 envelope_exponent=5):
-        super(SphericalBasisLayer, self).__init__()
-
-        assert num_radial <= 64
-        self.num_spherical = num_spherical
-        self.num_radial = num_radial
-        self.cutoff = cutoff
-        self.envelope = Envelope(envelope_exponent)
-
-        bessel_forms = bessel_basis(num_spherical, num_radial)
-        sph_harm_forms = real_sph_harm(num_spherical, zero_m_only = True)
-        self.sph_funcs = []
-        self.bessel_funcs = []
-
-        x, theta = sym.symbols('x theta')
-        modules = {'sin': torch.sin, 'cos': torch.cos}
-        for i in range(num_spherical):
-            if i == 0:
-                sph1 = sym.lambdify([theta], sph_harm_forms[i][0], modules)(0)
-                self.sph_funcs.append(lambda x: torch.zeros_like(x) + sph1)
-            else:
-                sph = sym.lambdify([theta], sph_harm_forms[i][0], modules)
-                self.sph_funcs.append(sph)
-            for j in range(num_radial):
-                bessel = sym.lambdify([x], bessel_forms[i][j], modules)
-                self.bessel_funcs.append(bessel)
-
-        self.bessel_forms = bessel_forms  # saving this because it takes forever to compute
-
-    def forward(self, dist, angle, idx_kj):
-        dist = dist / self.cutoff
-        rbf = torch.stack([f(dist) for f in self.bessel_funcs], dim=1)
-        rbf = self.envelope(dist).unsqueeze(-1) * rbf
-
-        cbf = torch.stack([f(angle) for f in self.sph_funcs], dim=1)
-
-        n, k = self.num_spherical, self.num_radial
-        out = (rbf[idx_kj].view(-1, n, k) * cbf.view(-1, n, 1)).view(-1, n * k)
-        return out
-
-
-class TorsionalEmbedding(torch.nn.Module):
-    def __init__(self, num_spherical, num_radial, cutoff=5.0,
-                 envelope_exponent=5, bessel_forms=None):
-        super(TorsionalEmbedding, self).__init__()
-        assert num_radial <= 64
-        self.num_spherical = num_spherical  #
-        self.num_radial = num_radial
-        self.cutoff = cutoff
-        self.envelope = Envelope(envelope_exponent)
-        if bessel_forms is None:
-            bessel_forms = bessel_basis(num_spherical, num_radial)
-        #        bessel_forms = bessel_basis(num_spherical, num_radial)
-        sph_harm_forms = real_sph_harm(num_spherical, zero_m_only=False)
-        self.sph_funcs = []
-        self.bessel_funcs = []
-
-        x = sym.symbols('x')
-        theta = sym.symbols('theta')
-        phi = sym.symbols('phi')
-        modules = {'sin': torch.sin, 'cos': torch.cos}
-        for i in range(self.num_spherical):
-            if i == 0:
-                sph1 = sym.lambdify([theta, phi], sph_harm_forms[i][0], modules)
-                self.sph_funcs.append(lambda x, y: torch.zeros_like(x) + torch.zeros_like(y) + sph1(0, 0))  # torch.zeros_like(x) + torch.zeros_like(y)
-            else:
-                for k in range(-i, i + 1):
-                    sph = sym.lambdify([theta, phi], sph_harm_forms[i][k + i], modules)
-                    self.sph_funcs.append(sph)
-            for j in range(self.num_radial):
-                bessel = sym.lambdify([x], bessel_forms[i][j], modules)
-                self.bessel_funcs.append(bessel)
-
-    def forward(self, dist, angle, phi, idx_kj):
-        dist = dist / self.cutoff
-        rbf = torch.stack([f(dist) for f in self.bessel_funcs], dim=1)
-        rbf = self.envelope(dist).unsqueeze(-1) * rbf
-
-        cbf = torch.stack([f(angle, phi) for f in self.sph_funcs], dim=1)
-
-        n, k = self.num_spherical, self.num_radial
-        out = (rbf[idx_kj].view(-1, 1, n, k) * cbf.view(-1, n, n, 1)).view(-1, n * n * k)
-        return out
-
-
-class BesselBasisLayer(torch.nn.Module):
+class BesselBasisLayer(torch.nn.Module):  # todo note borrowed from DimeNet implementation
     def __init__(self, num_radial, cutoff=5.0, envelope_exponent=5):
         super(BesselBasisLayer, self).__init__()
         self.cutoff = cutoff
