--- reporting/csp/utils.py
@@ -1,121 +1,0 @@
-import numpy as np
-from _plotly_utils.colors import n_colors
-from plotly import graph_objects as go
-from plotly.subplots import make_subplots
-from plotly.express.colors import sample_colorscale
-
-
-def log_mini_csp_scores_distributions(config, wandb, generated_samples_dict, real_samples_dict, real_data, sym_info):
-    """
-    report on key metrics from mini-csp
-    """
-    scores_labels = ['score', 'vdw overlap', 'density']  # , 'h bond score']
-    fig = make_subplots(rows=1, cols=len(scores_labels),
-                        vertical_spacing=0.075, horizontal_spacing=0.075)
-
-    colors = sample_colorscale('viridis', 1 + len(np.unique(generated_samples_dict['space group'])))
-    real_color = 'rgb(250,0,250)'
-    opacity = 0.65
-
-    for i, label in enumerate(scores_labels):
-        row = 1  # i // 2 + 1
-        col = i + 1  # i % 2 + 1
-        for j in range(min(15, real_data.num_graphs)):
-            bandwidth1 = np.ptp(generated_samples_dict[label][j]) / 50
-            real_score = real_samples_dict[label][j]
-
-            unique_space_group_inds = np.unique(generated_samples_dict['space group'][j])
-            n_space_groups = len(unique_space_group_inds)
-            space_groups = np.asarray([sym_info['space_groups'][sg] for sg in generated_samples_dict['space group'][j]])
-            unique_space_groups = np.asarray([sym_info['space_groups'][sg] for sg in unique_space_group_inds])
-
-            all_sample_score = generated_samples_dict[label][j]
-            for k in range(n_space_groups):
-                sample_score = all_sample_score[space_groups == unique_space_groups[k]]
-
-                fig.add_trace(go.Violin(x=sample_score, y=[str(real_data.csd_identifier[j]) for _ in range(len(sample_score))],
-                                        side='positive', orientation='h', width=2, line_color=colors[k],
-                                        meanline_visible=True, bandwidth=bandwidth1, opacity=opacity,
-                                        name=unique_space_groups[k], legendgroup=unique_space_groups[k], showlegend=False),
-                              row=row, col=col)
-
-            fig.add_trace(go.Violin(x=[real_score], y=[str(real_data.csd_identifier[j])], line_color=real_color,
-                                    side='positive', orientation='h', width=2, meanline_visible=True,
-                                    name="Experiment", showlegend=True if (i == 0 and j == 0) else False),
-                          row=row, col=col)
-
-            fig.update_xaxes(title_text=label, row=1, col=col)
-
-        unique_space_group_inds = np.unique(generated_samples_dict['space group'].flatten())
-        n_space_groups = len(unique_space_group_inds)
-        space_groups = np.asarray([sym_info['space_groups'][sg] for sg in generated_samples_dict['space group'].flatten()])
-        unique_space_groups = np.asarray([sym_info['space_groups'][sg] for sg in unique_space_group_inds])
-
-        if real_data.num_graphs > 1:
-            for k in range(n_space_groups):
-                all_sample_score = generated_samples_dict[label].flatten()[space_groups == unique_space_groups[k]]
-
-                fig.add_trace(go.Violin(x=all_sample_score, y=['all samples' for _ in range(len(all_sample_score))],
-                                        side='positive', orientation='h', width=2, line_color=colors[k],
-                                        meanline_visible=True, bandwidth=np.ptp(generated_samples_dict[label].flatten()) / 100, opacity=opacity,
-                                        name=unique_space_groups[k], legendgroup=unique_space_groups[k], showlegend=True if i == 0 else False),
-                              row=row, col=col)
-
-    layout = go.Layout(
-        margin=go.layout.Margin(
-            l=0,  # left margin
-            r=0,  # right margin
-            b=0,  # bottom margin
-            t=100,  # top margin
-        )
-    )
-    fig.update_xaxes(row=1, col=scores_labels.index('vdw overlap') + 1, range=[0, np.minimum(1, generated_samples_dict['vdw overlap'].flatten().max())])
-
-    fig.update_layout(yaxis_showgrid=True)  # legend_traceorder='reversed',
-
-    fig.layout.margin = layout.margin
-
-    if config.logger.log_figures:
-        wandb.log({'Mini-CSP Scores': fig})
-    if (config.machine == 'local') and False:
-        fig.show()
-
-    return None
-
-
-def log_csp_cell_params(config, wandb, generated_samples_dict, real_samples_dict, crystal_name, crystal_ind):
-    fig = make_subplots(rows=4, cols=3, subplot_titles=config.dataDims['lattice_features'])
-    for i in range(12):
-        bandwidth = np.ptp(generated_samples_dict['cell params'][crystal_ind, :, i]) / 100
-        col = i % 3 + 1
-        row = i // 3 + 1
-        fig.add_trace(go.Violin(
-            x=[real_samples_dict['cell params'][crystal_ind, i]],
-            bandwidth=bandwidth,
-            name="Samples",
-            showlegend=False,
-            line_color='darkorchid',
-            side='positive',
-            orientation='h',
-            width=2,
-        ), row=row, col=col)
-        for cc, cutoff in enumerate([0, 0.5, 0.95]):
-            colors = n_colors('rgb(250,50,5)', 'rgb(5,120,200)', 3, colortype='rgb')
-            good_inds = np.argwhere(generated_samples_dict['score'][crystal_ind] > np.quantile(generated_samples_dict['score'][crystal_ind], cutoff))[:, 0]
-            fig.add_trace(go.Violin(
-                x=generated_samples_dict['cell params'][crystal_ind, :, i][good_inds],
-                bandwidth=bandwidth,
-                name="Samples",
-                showlegend=False,
-                line_color=colors[cc],
-                side='positive',
-                orientation='h',
-                width=2,
-            ), row=row, col=col)
-
-    fig.update_layout(barmode='overlay', paper_bgcolor='rgba(0,0,0,0)', plot_bgcolor='rgba(0,0,0,0)')
-    fig.update_traces(opacity=0.5)
-    fig.update_layout(title=crystal_name)
-
-    wandb.log({"Mini-CSP Cell Parameters": fig})
-    return None
