+++ reporting/old_figures/nov_22_discriminator.py
