+++ reporting/old_figures/nov_22_discriminator_final.py
